+++ src/parser/macros.js
@@ -1,4 +1,5 @@
-import {Macro, Token} from "../classes.js";
+import {Macro} from "../macro.js"
+import {Token} from "../token.js"
 
 /**
  * Get macro by name. Name does not include `:` before the macro call
