+++ src/parser/parser.js
@@ -1,5 +1,5 @@
 import {DEFAULT_MACROS} from "../constants.js";
-import {Error} from "../classes.js";
+import {Error} from "../error.js";
 import {next, remaining, st, stn, update_src} from "./util.js";
 import {convertReservedChar, parse_inner, parseAttrs, parseComment, parseStr, parseTag} from "./main.js";
 import {handleImport, import_parse} from "./imports.js";
